common: move the rest of badblock_ndctl.c to libpmem2
